Move HOD classes to hod_models from hod
